% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psCompare.R
\name{ps.compare}
\alias{ps.compare}
\title{Propensity Score Comparison}
\usage{
ps.compare(data)
}
\arguments{
\item{data}{Data Frame - containing the dataset with previously calculated PS. The data
frame must contain a treatment indicator variable called 'treat' and a propensity score value
called 'ps_values'.}
}
\description{
Generates a graph used to compare the distribution of propensity score values in
the cotnrol and treatment groups
}
\details{
This function uses sm.density.compare to plot an overlapping density distribution plot
of the propensity scores for the two groups allowing one to visually assess the area
of common support
}
\examples{
ps.compare(myData)
}
