% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psOutcome.R
\name{ps.compute.survival}
\alias{ps.compute.survival}
\title{Compute Survival Objects}
\usage{
ps.compute.survival(data, outcome.dates, index.dates,
  obs.end.dates = "OBSEND", death.dates = "DEATHDATE")
}
\arguments{
\item{data}{Data frame, containing the dataset to be analyzed}

\item{outcome.dates}{String, containing the variable name with outcome dates}

\item{index.dates}{String, containing the variable name for the index dates}

\item{obs.end.dates}{String, containing the variable name for observation end dates (default = "OBSEND")}

\item{death.dates}{String, containing the variable name for death dates (default = "DEATHDATE")}
}
\value{
Vector, containing computed survival objects based on the outcome dates specified
}
\description{
Compute survival information using the survival package to be used in a response model
}
\details{
This function leverages the survival package Surv function to compute survival information to be
used as an outcome variable in outcome analysis. See survival::Surv documentation for more information.
}
\examples{
\dontrun{
ps.compute.survival(myData, "outcome")
}
}
