% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psDataConnector.R
\name{ps.connect.to.netezza}
\alias{ps.connect.to.netezza}
\title{Connect to Netezza}
\usage{
ps.connect.to.netezza(user.name, password, driver.location, db.prefix, db.host)
}
\arguments{
\item{user.name}{Your user name (prid)}

\item{password}{Your password}

\item{driver.location}{Location of your netezza driver, usually "nzjdbc.jar". Can be specified as complete
path or the directory where the driver file is located.}

\item{db.prefix}{This + prid = your database}

\item{db.host}{URL of netezza appliance}
}
\value{
Database connection object
}
\description{
Establishes a connection to the Netezza database
}
\details{
This function creates a connection to the Netezza database using the supplied credentials. This connection
can then be used to pull data from the database into R.
}
\examples{
ps.connect.to.netezza("user123", "password123", "/users/user123/Documents/databaseDrivers")
}
