% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psDiagnostics.R
\name{ps.covariate.biasplot}
\alias{ps.covariate.biasplot}
\title{Covariate Standard Differences of Means Plot}
\usage{
ps.covariate.biasplot(balanced, unbalanced, abs = FALSE)
}
\arguments{
\item{balanced}{Matrix, containing the summary statistics for covariates in the balanced population}

\item{unbalanced}{Matrix, containing the summary statistics for covariates in the unbalanced
/ total population}

\item{abs}{Boolean, indicating if the absolute percent reduction should be calculated. Default (FALSE)}
}
\description{
Visualization of the covariate standardized differences of means
}
\details{
This function creates a plot of each covariate's standardized differences of means before and after balancing
}
\examples{
ps.covariate.biasplot(summary.balanced, summary.unbalanced)
ps.covariate.biasplot(summary.balanced, summary.unbalanced, abs = TRUE)
}
