% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psDiagnostics.R
\name{ps.covariate.qq}
\alias{ps.covariate.qq}
\title{QQ Plots}
\usage{
ps.covariate.qq(data, covariates, weights = NULL)
}
\arguments{
\item{data}{Data frame, containing the balanced population data. The data
frame must contain a treatment indicator variable called 'treat'.}

\item{covariates}{Vector, containing the list of covariate names}

\item{weights}{Vector, containing the weights to use in balanced population calculations. Defaults to unweighted.}
}
\description{
Create QQ plots for each of the continuous covariates
}
\details{
This function identifies each of the continuous covariates and creates a QQ plot, which allows one
to visually inspect the similarity of distributions of covariates in the control and treatment populations
}
\examples{
\dontrun{
ps.covariate.qq(myData, covariates)
ps.covariate.qq(myData, covariates, myData$is_matched)
}
}
