% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psDiagnostics.R
\name{ps.covariate.residualplot}
\alias{ps.covariate.residualplot}
\title{Covariate Residuals Plot}
\usage{
ps.covariate.residualplot(data, covariates, weights = NULL,
  dichotomous = FALSE)
}
\arguments{
\item{data}{Data Frame, containing the dataset. The data
frame must contain a treatment indicator variable called 'treat'.}

\item{covariates}{Vector, containing the list of covariates to include in the plot}

\item{weights}{Vector, containing the weights to use in assessment of balanced population. Defaults to unweighted}

\item{dichotomous}{Boolean value indicating if dichotomous covariates should be included in the calculation or not}
}
\description{
A graphical balance diagnostic examining the ratio of residuals orthogonal to the propensity scores
}
\details{
This function creates a plot of the residuals orthogonal to the propensity scores. Each covariate is
regressed on the propensity scores for the control and treatment populations. The variance of the residuals
of this regression are then compared. For a balanced population, the ratio of variances should be close to
1. An imbalance in the variances indicates that there is additional variance potentially attributed to the
covariate in one of the populations that has not been captured in the propensity scores.
}
\examples{
ps.covariate.residualplot(myData, covariates)
}
