% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psDiagnostics.R
\name{ps.covariate.statistics}
\alias{ps.covariate.statistics}
\title{Covariate Summary Matrix}
\usage{
ps.covariate.statistics(data, covariates, weights = NULL, outputs = c("all",
  "dichotomous", "continuous"))
}
\arguments{
\item{data}{Data Frame - containing the dataset with previously calculated weights or matches. The data
frame must contain a treatment indicator variable called 'treat'.}

\item{covariates}{Vector - containing the set of covariates for which to evaluate statistics.}

\item{weights}{Vector, containing the weights to use in evaluating statistics for population. Defaults to
equal weighting.}

\item{outputs}{String - indicates which covariate type(s) to output. Accepted inputs are "all" (Default), "dichotomous", and "continuous"}
}
\value{
Matrix - containing summary statistics for covariates
}
\description{
Calculate the summary statistics for covariates
}
\details{
This function calculates a variety of summary statistics and organizes them into a matrix.
Included summary statistics are the mean and sd in the treatment and control populations,
the difference in means between the treat and control populations and the standardized
difference of the means between the populations.
}
\examples{
\dontrun{
ps.covariate.statistics(myData, covariates)
ps.covariate.statistics(myData, covariates, weights = myData$weights)
ps.covariate.statistics(myData, covariates, weights = myData$is_matched)
ps.covariate.statistics(myData, covariates, outputs = "dichotomous")
}
}
