% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psCohortMethodData.R
\name{ps.getCohortMethodData}
\alias{ps.getCohortMethodData}
\title{Cohort Method Data}
\usage{
ps.getCohortMethodData(data, outcomes, covariates, treat.name = "treat",
  person.id.name = "PERSON_ID")
}
\arguments{
\item{data}{Data Frame - containing the dataset to reformat.}

\item{outcomes}{- outcome vector, or list of outcome vectors to include in data object. For TTE outcomes,
include event times (NA for no event), for dichtomous data include 1/0 indicator}

\item{covariates}{Vector of covariate names to include in the data object}

\item{treat.name}{Name of the treatment indicator [0,1] variable (default "treat")}

\item{person.id.name}{Name of the subject id variable (default "PERSON_ID")}
}
\value{
CohortMethodData - reformatted data object
}
\description{
Reformat a data frame into a CohortMethodData object for use with OHDSI package
}
\details{
This function takes a data frame containing an observational study dataset and reformats
into an object of type CohortMethodData that can be used with the OHDSI CohortMethod library.
}
\examples{
\dontrun{
ps.getCohortMethodData(myData)
}
}
