% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psOutcome.R
\name{ps.km.plot}
\alias{ps.km.plot}
\title{Kaplan-Meier Plot}
\usage{
ps.km.plot(km.analysis, title = "Kaplan-Meier Curves")
}
\arguments{
\item{km.analysis}{A survival analysis object. See ps.survival()}

\item{title}{Chart title (default "Kaplan-Meier Curves")}
}
\description{
Creates a basic Kaplan-Meier plot for the survival analysis
}
\details{
This function creates a basic Kaplan-Meier plot for the supplied survival analysis. Basic assumptions are
included about extent of axes, labels, etc.
}
\examples{
ps.km.plot(model$km)
}
