% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psDiagnostics.R
\name{ps.matched.boxplot}
\alias{ps.matched.boxplot}
\title{Matched Sample Box and Whisker}
\usage{
ps.matched.boxplot(data)
}
\arguments{
\item{data}{Data Frame - containing the dataset with previously calculated matches. The data
frame must contain a treatment indicator variable called 'treat' and a variable indicating whether
the subject was matched called 'is_matched'.}
}
\description{
Simple visual representation of matched samples
}
\details{
This function creates a box and whisker plot of matched and unmatched samples.
}
\examples{
ps.matched.boxplot(myData)
}
