% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psDiagnostics.R
\name{ps.matched.scatterplot}
\alias{ps.matched.scatterplot}
\title{Matched Sample Scatterplot}
\usage{
ps.matched.scatterplot(data)
}
\arguments{
\item{data}{Data Frame - containing the dataset with previously calculated matches.  The data
frame must contain a treatment indicator variable called 'treat' and a variable indicating whether
the subject was matched called 'is_matched'.}
}
\description{
Simple visual representation of matched samples
}
\details{
This function creates a simple scatter plot of matched and unmatched samples. Works best for
smaller sample sizes.
}
\examples{
ps.matched.scatterplot(myData)
}
