% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psDiagnostics.R
\name{ps.matched.summary}
\alias{ps.matched.summary}
\title{Simple population summary of matched samples}
\usage{
ps.matched.summary(data)
}
\arguments{
\item{data}{Data Frame - containing the dataset with previously calculated is_matched. The data
frame must contain a treatment indicator variable called 'treat' and a variable indicating whether
the subject was matched called 'is_matched'.}
}
\value{
Matrix - containing the summary counts of treat and control in original and matched
               populations
}
\description{
Summarize the matched samples.
}
\details{
This function does a simple summary counting of the matched vs unmatched samples.
Print the resulting table to see counts of the original population, vs the
populations that result from previously completed matching. (Not really useful
for weighting as all subjects will have a non-zero weight)
}
\examples{
\dontrun{
ps.matched.summary(myData)
}
}
