% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psDiagnostics.R
\name{ps.matched.violinplot}
\alias{ps.matched.violinplot}
\title{Matched Sample Violinplot}
\usage{
ps.matched.violinplot(data)
}
\arguments{
\item{data}{Data Frame - containing the dataset with previously calculated matches. The data
frame must contain a treatment indicator variable called 'treat' and a variable indicating whether
the subject was matched called 'is_matched'.}
}
\description{
Simple visual representation of matched samples
}
\details{
This function creates a violin plot of matched and unmatched samples. This plot shows the
kernel density for the two populations - with an inner box and whisker plot of summary statistics.
This plot tends to work better for large populations than the matched scatter plot.
}
\examples{
\dontrun{
ps.matched.violinplot(myData)
}
}
