% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psDiagnostics.R
\name{ps.percent.reduction}
\alias{ps.percent.reduction}
\title{Percent Reduction in Standardized Differences of Means}
\usage{
ps.percent.reduction(balanced, unbalanced, abs = FALSE)
}
\arguments{
\item{balanced}{Matrix, containing the summary statistics for covariates in the balanced population}

\item{unbalanced}{Matrix, containing the summary statistics for covariates in the unbalanced
/ total population}

\item{abs}{Boolean, indicating if the absolute percent reduction should be calculated. Default (FALSE)}
}
\value{
Matrix, containing the percent reduction in standardized differences of means for all covariates
}
\description{
Calculate the percent reduction in the standardized differences of the means for
covariates
}
\details{
This function uses the calculated covariate summary information for a balanced and
total / unbalanced population and computes the percent reduction in the standardized
differences of the means
}
\examples{
ps.percent.reduction(summary.balanced, summary.unbalanced)
}
