% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psAUC.R
\name{ps.roc}
\alias{ps.roc}
\title{ROC}
\usage{
ps.roc(data)
}
\arguments{
\item{data}{Data Frame - containing the dataset with previously calculated PS. The data
frame must contain a treatment indicator variable called 'treat' and a propensity score value
called 'ps_values'.}
}
\description{
Generates a ROC curve for a set of calculated propensity scores.
}
\details{
This function uses the 'AUC' R package to generate a ROC curve of the
propensity scores. For this to work properly, the argument \code{...} should
be a data frame containing the dataset with previously calculated propensity
scores in a variable called ps_values.
}
\examples{
\dontrun{
ps.roc(myData)
}
}
