% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psDiagnostics.R
\name{ps.score.summary}
\alias{ps.score.summary}
\title{Propensity Score Balance}
\usage{
ps.score.summary(data, weights = NULL)
}
\arguments{
\item{data}{Data frame, containing the balanced population, including weights and/or matches. The data
frame must contain a treatment indicator variable called 'treat'.}

\item{weights}{Vector, containing the weights to use in assessing the balanced population (defaults to unweighted).}
}
\value{
Matrix, containing the summary statistics for the calculated propensity score
}
\description{
Summarizes the balance of the propensity scores in treatment and control populations
}
\details{
This function calculates key summary statistics based on the propensity scores in the treatment
and control populations. For a balanced population, one should observe similar mean values of the
propensity scores in the two populations. Similarly to covariate diagnostics, one should observe a
decrease in the standardized difference of the mean propensity score after balancing. Finally, one
should see the ratio of variances in the propensity score move closer to 1 after balancing.
}
\examples{
ps.score.summary(myData)
ps.score.summary(myData, weights = myData$is_matched)
}
