% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psAUC.R
\name{ps.sensitivity}
\alias{ps.sensitivity}
\title{Sensitivity}
\usage{
ps.sensitivity(data)
}
\arguments{
\item{data}{Data Frame - containing the dataset with previously calculated PS. The data
frame must contain a treatment indicator variable called 'treat' and a propensity score value
called 'ps_values'.}
}
\value{
Calculated sensitivity of propensity scores
}
\description{
Calculates the sensitivity curve of calculated propensity scores.
}
\details{
This function uses the 'AUC' R package to calculate the sensitivity of the
propensity scores. For this to work properly, the argument \code{...} should
be a data frame containing the dataset with previously calculated propensity
scores in a variable called ps_values.
}
\examples{
\dontrun{
ps.sensitivity(myData)}
}
