% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psDataConnector.R
\name{ps.sql2df}
\alias{ps.sql2df}
\title{SQL to Data.Frame}
\usage{
ps.sql2df(conn, sql.query)
}
\arguments{
\item{conn}{Database connection object created by ps.connect.to.netezza}

\item{sql.query}{SQL query string, just as you might type into Aginity}
}
\value{
Calculated accuracy of propensity scores
}
\description{
Execute a sql query and return the results as a data.frame
}
\details{
This function uses a database connection to execute a sql query string. The results returned by the query
are returned as an R data frame.
}
\examples{
myData <- ps.sql2df(nzconn, "select * from database.table limit 100")
}
