% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psStrata.R
\name{ps.strata.balance}
\alias{ps.strata.balance}
\title{Strata Balance}
\usage{
ps.strata.balance(data, strata, covariates)
}
\arguments{
\item{data}{Data Frame, containing the dataset. The data
frame must contain a treatment indicator variable called 'treat'.}

\item{strata}{Vector, containing the strata for the dataset}

\item{covariates}{Vector, containing the covariate variable names to include}
}
\description{
Visualize the covariate balance within each stratum
}
\details{
This function creates a bar or box-and-whisker plot visualizing the summary statistics
for each covariate in each of the strata
}
\examples{
\dontrun{
ps.strata.balance(myData, myData$strata, covariates)
}
}
