% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psStrata.R
\name{ps.strata.survival}
\alias{ps.strata.survival}
\title{Evaluate a survival model}
\usage{
ps.strata.survival(data, strata, outcome, estimand = "ATT",
  covariates = character())
}
\arguments{
\item{data}{Data Frame - containing the stratified dataset. The data
frame must contain a treatment indicator variable called 'treat' and a propensity score value
called 'ps_values'.}

\item{strata}{Vector - containing the stratum that each subject is assigned to}

\item{outcome}{The outcome variable name. The variable must contain a precomputed survival object}

\item{estimand}{String, the desired estimand ("ATT" or "ATE"), default (ATT).}

\item{covariates}{Vector, containing the names of the additional covariates to include in the outcome analysis}
}
\value{
An object of type "strata.model". This object contains the model information for each strata in object$models,
         as well as the pooled treatment effect information in object$treatment.effect.
}
\description{
Evaluate a survival outcome model for the stratified  dataset.
}
\details{
This function evaluates an outcome model for each stratum and then pools the models
to evaluate the overall treatment effect
}
\examples{
\dontrun{
ps.strata.survival(myData, myData$strata, "MISURVIVAL")
}
}
