% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psStrata.R
\name{ps.stratify}
\alias{ps.stratify}
\title{Compute Strata}
\usage{
ps.stratify(data, n = 5)
}
\arguments{
\item{data}{Data Frame - containing the dataset with previously calculated PS. The data frame
must contain propensity scores in a variable called 'ps_values'.}

\item{n}{Number of strata to divide the dataset into (default 5)}
}
\value{
Vector - containing the categorical strata number for each subject
}
\description{
Stratifies the popultion.
}
\details{
This function stratifies a popultion based on an estimated propensity score. The stratum
of each subject is added to the dataset as a categorical value.
}
\examples{
\dontrun{
ps.stratify(myData)
ps.stratify(myData, n = 10)
}
}
