% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psOutcome.R
\name{ps.survival}
\alias{ps.survival}
\title{Outcome Survival Model}
\usage{
ps.survival(data, outcome, covariates = character(), w = NULL)
}
\arguments{
\item{data}{Data frame, containing the dataset to be analyzed. The data
frame must contain a treatment indicator variable called 'treat' and propensity score values called 'ps_values'.}

\item{outcome}{String, containing the outcome variable name to be analyzed. The variable must be a Surv object}

\item{covariates}{Vector, containing the set of covariate variable names to include in the model}

\item{w}{Vector, containing the subject weights. Defaults to equal weighting.
If analysis of matched data is desired, set this value to myData$is_matched.}
}
\value{
Object, containing fitted model values. In addition to standard glm/lm output, the treatment effect
         is appended to the model object as model$treatment.effect. For dichotomous outcome variables, this is
         the odds ratio with confidence interval.
}
\description{
Performs survival analysis of an outcome variable
}
\details{
This function uses the survival package to do basic survival analysis for an outcome variable. The outcome
variable should be previously computed as a survival object (see ps.compute.survival). The analysis will
automatically perform survival::survfit and survival::coxph.
}
\examples{
ps.survival(myData, "outcome")
}
