% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psTrim.R
\name{ps.trim}
\alias{ps.trim}
\title{Trim Population}
\usage{
ps.trim(data, trim.method = "overlap", trim.quantile = 0.95,
  quantile.group = "all")
}
\arguments{
\item{data}{Data Frame - containing the dataset with previously calculated PS. The data
frame must contain a treatment indicator variable called 'treat' and propensity score values called 'ps_values'.}

\item{trim.method}{String, specifying the method to use to trim the dataset. Available options
are "overlap" (default) to trim the non-overlapping tails of the control and
treatment distributions, or "quantile" to trim to an interquantil range.}

\item{trim.quantile}{Number, specifying the interquantile range to be trimmed. For example,
0.95 indicates that the dataset should be trimmed to the 0.025 / 0.975 interquantile range.}

\item{quantile.group}{Specifies the group on which to determine the quantile values. Options are "all" (default)
to determine the quantile values based on the propensity scores for the entire population, or
"treat" to determine the quantile values based on the propensity scores for only the treatment
group}
}
\value{
Data Frame - trimmed dataset
}
\description{
Trims the population based on the propensity scores
}
\details{
This function trims the supplied population based on propensity scores. Trimming can be
based on the overlapping regions, or a specified interquantile range.
}
\examples{
ps.trim(myData)
ps.trim(myData, trim.method = "quantile", trime.quantile = 0.90)
}
